recko_backend/requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
attrs==20.3.0
autopep8==1.5.5
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.6
cached-property==1.5.1
cachelib==0.1.1
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
cryptography==3.2.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==2.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-cursor-pagination==0.1.4
django-debug-toolbar==3.2
django-filter==2.4.0
django-postgres-fuzzycount==0.1.7
django-progressive-web-app==0.1.1
django-pwa==1.0.10
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-registration==0.5.6
django-rest-swagger==2.2.0
django-social-auth==0.7.28
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
ecdsa==0.14.1
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.10
eventlet==0.21.0
extras==1.0.0
fanstatic==1.1
fasteners==0.14.1
filelock==3.0.12
fixtures==3.0.0
Flask==1.1.2
Flask-Session==0.3.2
future==0.18.2
google-auth==1.4.2
greenlet==1.0.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
intuit-oauth==1.1.1
ipaddr==2.2.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
js.jquery==3.3.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.2.0
language-check==1.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==5.2.2
npm==0.1.1
numpy==1.19.1
oauth2==1.9.0.post1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
opencv-python==4.4.0.42
optional-django==0.1.0
packaging==20.3
pandas==1.1.1
pandocfilters==1.4.2
paramiko==2.6.0
pbr==5.5.0
pep517==0.8.2
pexpect==4.6.0
phylactery==0.2.2
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
portpicker==1.2.0
postgres==3.0.0
progress==1.5
prompt-toolkit==1.0.18
protobuf==3.6.1
psycopg2-binary==2.8.6
psycopg2-pool==1.1
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycosat==0.6.3
pycountry==18.12.8
pycparser==2.20
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.16.0
pytest==6.2.2
pytest-cov==2.5.0
python-apt==2.0.0+ubuntu0.20.4.4
python-coveralls==2.9.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
python-jose==3.2.0
python-mimeparse==1.6.0
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyxdg==0.26
pyxero==0.9.2
PyYAML==5.4.1
pyzmq==19.0.2
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schedule==1.0.0
SecretStorage==2.3.1
shutilwhich==1.1.0
simplegeneric==0.8.1
simplejson==3.17.2
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sphinx-me==0.3
SQLAlchemy==1.3.19
sqlparse==0.4.1
stevedore==3.2.2
systemd-python==234
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
testtools==2.4.0
texttable==1.6.2
tld==0.12.5
toml==0.10.2
torch-transformers==1.0.4
tornado==4.5.3
traceback2==1.4.0
traitlets==5.0.0
twilio==6.45.1
typing==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
ural==0.25.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
websocket-client==0.53.0
Werkzeug==1.0.1
whitenoise==5.2.0
xero-python==1.4.0
xkit==0.0.0
yapf==0.30.0
youtube-transcript-api==0.3.1
zipp==1.0.0
